requirements.txt
pymupdf==1.26.3
pillow==11.1.0
google-cloud-vision==3.10.2
